backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
alembic==1.13.1
asyncpg==0.29.0
pydantic==2.6.3
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pgvector==0.2.5
google-generativeai==0.4.1
python-dotenv==1.0.1
httpx==0.27.0
pypdf==4.1.0
langchain-text-splitters==0.0.1
moviepy==1.0.3
